# openapi.json
{"openapi": "3.1.0", "info": {"title": "FastAPI", "version": "0.1.0"}, "paths": {"/api/v1/webhook/incident": {"post": {"summary": "Webhook", "operationId": "webhook_api_v1_webhook_incident_post", "parameters": [{"name": "authorization", "in": "header", "required": false, "schema": {"type": "string", "title": "Authorization"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Payload"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/health": {"get": {"summary": "Health Check", "operationId": "health_check_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/ready": {"get": {"summary": "Readiness Check", "operationId": "readiness_check_ready_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/v1/config": {"get": {"summary": "Get Config", "operationId": "get_config_api_v1_config_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/v1/executions/{execution_id}": {"get": {"summary": "Get Execution", "operationId": "get_execution_api_v1_executions__execution_id__get", "parameters": [{"name": "execution_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Execution Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ExecutionResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/executions/{execution_id}/trace": {"get": {"summary": "Get Execution Trace", "operationId": "get_execution_trace_api_v1_executions__execution_id__trace_get", "parameters": [{"name": "execution_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Execution Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ExecutionTraceResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/incidents/{sys_id}/executions": {"get": {"summary": "List Executions For Incident", "operationId": "list_executions_for_incident_api_v1_incidents__sys_id__executions_get", "parameters": [{"name": "sys_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Sys Id"}}, {"name": "page", "in": "query", "required": false, "schema": {"type": "integer", "default": 1, "title": "Page"}}, {"name": "page_size", "in": "query", "required": false, "schema": {"type": "integer", "default": 20, "title": "Page Size"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PaginatedExecutionsResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/approvals": {"get": {"summary": "List Approvals", "operationId": "list_approvals_api_v1_approvals_get", "parameters": [{"name": "page", "in": "query", "required": false, "schema": {"type": "integer", "default": 1, "title": "Page"}}, {"name": "page_size", "in": "query", "required": false, "schema": {"type": "integer", "default": 20, "title": "Page Size"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ApprovalListResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/approvals/{approval_id}/decide": {"post": {"summary": "Decide Approval", "operationId": "decide_approval_api_v1_approvals__approval_id__decide_post", "parameters": [{"name": "approval_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Approval Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ApprovalDecision"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ApprovalDecisionResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/dlq": {"get": {"summary": "List Dlq", "operationId": "list_dlq_api_v1_dlq_get", "parameters": [{"name": "page", "in": "query", "required": false, "schema": {"type": "integer", "default": 1, "title": "Page"}}, {"name": "page_size", "in": "query", "required": false, "schema": {"type": "integer", "default": 20, "title": "Page Size"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DLQListResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/dlq/{event_id}/replay": {"post": {"summary": "Replay Dlq Event", "operationId": "replay_dlq_event_api_v1_dlq__event_id__replay_post", "parameters": [{"name": "event_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Event Id"}}, {"name": "authorization", "in": "header", "required": false, "schema": {"type": "string", "title": "Authorization"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DLQReplayResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/eval/results": {"get": {"summary": "Get Eval Results", "operationId": "get_eval_results_api_v1_eval_results_get", "parameters": [{"name": "page", "in": "query", "required": false, "schema": {"type": "integer", "default": 1, "title": "Page"}}, {"name": "page_size", "in": "query", "required": false, "schema": {"type": "integer", "default": 20, "title": "Page Size"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/EvalResultsListResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/eval/run": {"post": {"summary": "Trigger Eval Run", "operationId": "trigger_eval_run_api_v1_eval_run_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/EvalRunRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/EvalRunResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"ApprovalDecision": {"properties": {"action": {"type": "string", "title": "Action"}, "reviewer": {"type": "string", "title": "Reviewer"}, "rationale": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Rationale"}}, "type": "object", "required": ["action", "reviewer"], "title": "ApprovalDecision"}, "ApprovalDecisionResponse": {"properties": {"approval_id": {"type": "string", "title": "Approval Id"}, "status": {"type": "string", "title": "Status"}, "reviewer": {"type": "string", "title": "Reviewer"}, "decided_at": {"type": "string", "format": "date-time", "title": "Decided At"}}, "type": "object", "required": ["approval_id", "status", "reviewer", "decided_at"], "title": "ApprovalDecisionResponse"}, "ApprovalListResponse": {"properties": {"items": {"items": {"$ref": "#/components/schemas/ApprovalResponse"}, "type": "array", "title": "Items"}, "page": {"type": "integer", "title": "Page"}, "page_size": {"type": "integer", "title": "Page Size"}, "total": {"type": "integer", "title": "Total"}}, "type": "object", "required": ["items", "page", "page_size", "total"], "title": "ApprovalListResponse"}, "ApprovalResponse": {"properties": {"approval_id": {"type": "string", "title": "Approval Id"}, "incident_sys_id": {"type": "string", "title": "Incident Sys Id"}, "status": {"type": "string", "title": "Status"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}}, "type": "object", "required": ["approval_id", "incident_sys_id", "status", "created_at"], "title": "ApprovalResponse"}, "DLQEntryResponse": {"properties": {"event_id": {"type": "string", "title": "Event Id"}, "original_payload": {"additionalProperties": true, "type": "object", "title": "Original Payload"}, "failure_reason": {"type": "string", "title": "Failure Reason"}, "retry_count": {"type": "integer", "title": "Retry Count"}, "timestamp": {"type": "string", "format": "date-time", "title": "Timestamp"}}, "type": "object", "required": ["event_id", "original_payload", "failure_reason", "retry_count", "timestamp"], "title": "DLQEntryResponse"}, "DLQListResponse": {"properties": {"items": {"items": {"$ref": "#/components/schemas/DLQEntryResponse"}, "type": "array", "title": "Items"}, "page": {"type": "integer", "title": "Page"}, "page_size": {"type": "integer", "title": "Page Size"}, "total": {"type": "integer", "title": "Total"}}, "type": "object", "required": ["items", "page", "page_size", "total"], "title": "DLQListResponse"}, "DLQReplayResponse": {"properties": {"event_id": {"type": "string", "title": "Event Id"}, "status": {"type": "string", "title": "Status"}, "requeued_at": {"type": "string", "format": "date-time", "title": "Requeued At"}}, "type": "object", "required": ["event_id", "status", "requeued_at"], "title": "DLQReplayResponse"}, "EvalResultResponse": {"properties": {"run_id": {"type": "string", "title": "Run Id"}, "status": {"type": "string", "title": "Status"}, "metrics": {"additionalProperties": true, "type": "object", "title": "Metrics"}, "total_samples": {"type": "integer", "title": "Total Samples"}, "completed_at": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Completed At"}}, "type": "object", "required": ["run_id", "status", "metrics", "total_samples"], "title": "EvalResultResponse"}, "EvalResultsListResponse": {"properties": {"items": {"items": {"$ref": "#/components/schemas/EvalResultResponse"}, "type": "array", "title": "Items"}, "page": {"type": "integer", "title": "Page"}, "page_size": {"type": "integer", "title": "Page Size"}, "total": {"type": "integer", "title": "Total"}}, "type": "object", "required": ["items", "page", "page_size", "total"], "title": "EvalResultsListResponse"}, "EvalRunRequest": {"properties": {"dataset_id": {"type": "string", "title": "Dataset Id"}, "agent_version": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Agent Version"}, "model": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Model"}, "parameters": {"anyOf": [{"additionalProperties": true, "type": "object"}, {"type": "null"}], "title": "Parameters"}}, "type": "object", "required": ["dataset_id"], "title": "EvalRunRequest"}, "EvalRunResponse": {"properties": {"run_id": {"type": "string", "title": "Run Id"}, "status": {"type": "string", "title": "Status"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}}, "type": "object", "required": ["run_id", "status", "created_at"], "title": "EvalRunResponse"}, "ExecutionResponse": {"properties": {"execution_id": {"type": "string", "title": "Execution Id"}, "incident_sys_id": {"type": "string", "title": "Incident Sys Id"}, "status": {"type": "string", "title": "Status"}, "current_node": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Current Node"}, "started_at": {"type": "string", "format": "date-time", "title": "Started At"}, "completed_at": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Completed At"}, "model": {"type": "string", "title": "Model"}, "agent_version": {"type": "string", "title": "Agent Version"}}, "type": "object", "required": ["execution_id", "incident_sys_id", "status", "started_at", "model", "agent_version"], "title": "ExecutionResponse"}, "ExecutionTraceResponse": {"properties": {"execution_id": {"type": "string", "title": "Execution Id"}, "nodes": {"items": {"$ref": "#/components/schemas/TraceNode"}, "type": "array", "title": "Nodes"}}, "type": "object", "required": ["execution_id", "nodes"], "title": "ExecutionTraceResponse"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "PaginatedExecutionsResponse": {"properties": {"items": {"items": {"$ref": "#/components/schemas/ExecutionResponse"}, "type": "array", "title": "Items"}, "page": {"type": "integer", "title": "Page"}, "page_size": {"type": "integer", "title": "Page Size"}, "total": {"type": "integer", "title": "Total"}}, "type": "object", "required": ["items", "page", "page_size", "total"], "title": "PaginatedExecutionsResponse"}, "Payload": {"properties": {"event_id": {"type": "string", "title": "Event Id", "description": "The unique identifier for the webhook event"}, "sys_id": {"type": "string", "title": "Sys Id", "description": "The unique identifier for the incident"}, "number": {"type": "string", "title": "Number", "description": "The incident number"}, "event_type": {"type": "string", "title": "Event Type", "description": "The type of event that triggered the webhook"}, "contract_version": {"type": "string", "title": "Contract Version", "description": "The version of the contract that defines the webhook payload structure"}}, "type": "object", "required": ["event_id", "sys_id", "number", "event_type", "contract_version"], "title": "Payload"}, "TraceNode": {"properties": {"node_name": {"type": "string", "title": "Node Name"}, "entered_at": {"type": "string", "format": "date-time", "title": "Entered At"}, "output": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Output"}}, "type": "object", "required": ["node_name", "entered_at"], "title": "TraceNode"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}